Created folder for identity interfaces and moved them there.
