+++ FlowerLovers.Core/Contracts/IdentityServices/IPersonalDataService.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 
-namespace FlowerLovers.Core.Contracts
+namespace FlowerLovers.Core.Contracts.IdentityServices
 {
     public interface IPersonalDataService
     {
